Add FlexSearch plugin with search functionality for Nikola sites (#441)
This commit introduces a new FlexSearch plugin that provides search functionality for Nikola static sites. The plugin generates a JSON file containing article data, which is used with FlexSearch to enable search capabilities on the site. It includes instructions on how to implement search results in a div or an overlay. Also, contains added configuration options in a sample `conf.py.sample` file which allows users to customize according to their needs.